+++ dh-module-oa/dh-module-oa-server/src/main/java/cn/dh/oa/module/oa/service/document/DocumentDispatchBillServiceImpl.java
@@ -8,6 +8,7 @@ import cn.dh.oa.module.bpm.api.task.dto.BpmProcessInstanceCreateReqDTO;
 import cn.dh.oa.module.bpm.enums.task.BpmTaskStatusEnum;
 import cn.dh.oa.module.oa.enums.OaBillTypeEnum;
 import cn.dh.oa.module.oa.service.bill.OaBillApprovalVisibleService;
+import cn.dh.oa.module.bpm.util.BpmProcessInstanceCancelUtils;
 import cn.dh.oa.module.bpm.util.BpmProcessVariableUtils;
 import cn.dh.oa.framework.common.service.FlowBillService;
 import cn.dh.oa.common.server.attachment.service.AttachmentService;
@@ -17,6 +18,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import jakarta.annotation.Resource;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
 import cn.dh.oa.module.oa.controller.admin.document.vo.*;
@@ -129,16 +131,23 @@ public class DocumentDispatchBillServiceImpl implements DocumentDispatchBillServ
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public void deleteDocumentDispatchBill(Long id) {
-        // 校验存在
-        validateDocumentDispatchBillExists(id);
-        // 删除
+        DocumentDispatchBillDO bill = documentDispatchBillMapper.selectById(id);
+        if (bill == null) {
+            throw exception(DOCUMENT_DISPATCH_BILL_NOT_EXISTS);
+        }
+        BpmProcessInstanceCancelUtils.cancelIfExists(processInstanceApi, bill.getProcessInstanceId());
         documentDispatchBillMapper.deleteById(id);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public void deleteDocumentDispatchBillListByIds(List<Long> ids) {
-        // 删除
+        List<DocumentDispatchBillDO> bills = documentDispatchBillMapper.selectByIds(ids);
+        for (DocumentDispatchBillDO bill : bills) {
+            BpmProcessInstanceCancelUtils.cancelIfExists(processInstanceApi, bill.getProcessInstanceId());
+        }
         documentDispatchBillMapper.deleteByIds(ids);
     }
